refactor(cleanup): consolidate small guards and neutralize references
