+++ CHANGELOG.md
@@ -58,7 +58,7 @@ this project adheres to
 ### Added
 
 - **Missing CSS Module diagnostic** — `import styles from './typo.module.scss'` now emits a `missing-module` warning when the target file does not exist on disk. Fires for any file with a CSS Module import, including pure `styles.x` access without `classnames/bind`. Configurable via `cssModuleExplainer.diagnostics.missingModule` (default `true`).
-- **Path alias compat — `cssModules.pathAlias`** — the clinyong/vscode-cssmodules `cssModules.pathAlias` config is read as-is, so `import styles from '@styles/button.module.scss'` resolves when the workspace has `"cssModules.pathAlias": { "@styles": "src/styles" }` in its settings. Zero-config migration for clinyong users. **One intentional divergence**: we use longest-prefix matching instead of clinyong's insertion-order first-match, so `{ "@": "src", "@styles": "src/styles" }` correctly routes `@styles/button` to `src/styles/button` regardless of config key order. `${workspaceFolder}` substitution is supported. Wildcards and tsconfig.json `compilerOptions.paths` auto-detection are not yet supported — tracked for a future release.
+- **Path alias compat — `cssModules.pathAlias`** — the `cssModules.pathAlias` config is read as-is, so `import styles from '@styles/button.module.scss'` resolves when the workspace has `"cssModules.pathAlias": { "@styles": "src/styles" }` in its settings. Existing workspace settings continue to work without migration. Alias matching uses longest-prefix order, so `{ "@": "src", "@styles": "src/styles" }` correctly routes `@styles/button` to `src/styles/button` regardless of config key order. `${workspaceFolder}` substitution is supported. Wildcards and tsconfig.json `compilerOptions.paths` auto-detection are not yet supported — tracked for a future release.
 - **`classnameTransform` setting** — expose SCSS classes under five modes matching css-loader's `localsConvention`: `asIs` (default, unchanged behavior), `camelCase` (original + camelCase alias), `camelCaseOnly` (camelCase only), `dashes` (original + dashes-to-camel alias), `dashesOnly` (dashes-to-camel only). With `camelCase` active, both `styles['btn-primary']` and `styles.btnPrimary` resolve against a single `.btn-primary` selector. Alias entries participate in BEM-safe rename — renaming `btnPrimary` rewrites the original `.btn-primary` token in SCSS and every call site in TSX, with each site getting the form that matches how it accesses the class. `camelCaseOnly` and `dashesOnly` reject alias rename because the reverse transform from camelCase back to the original SCSS separator is lossy; use `camelCase` / `dashes` for editor-driven rename workflows. Configurable via `cssModuleExplainer.scss.classnameTransform` (default `"asIs"`).
 
 ## [1.5.1] — 2026-04-11
@@ -85,7 +85,7 @@ this project adheres to
 ### Changed
 
 - **Unified `ClassRef` domain model** — legacy `CxCallInfo` and `StylePropertyRef` types collapsed into a single `ClassRef` discriminated union (`static | template | variable`, tagged with `origin: "cxCall" | "styleAccess"`). Every provider now dispatches through a single `withClassRefAtCursor` front stage; the parallel `withCxCallAtCursor` / `withStyleRefAtCursor` dispatch pattern is gone.
-- **Error boundary at every LSP handler** — new `wrapHandler(name, impl, fallback)` helper (Stylable-inspired) wraps each provider export with a try/catch + `logError` + safe default. The nine hand-rolled try/catch blocks in individual providers are deleted.
+- **Error boundary at every LSP handler** — new `wrapHandler(name, impl, fallback)` helper wraps each provider export with a try/catch + `logError` + safe default. The nine hand-rolled try/catch blocks in individual providers are deleted.
 - **Single-source DI** — `HandlerContext.getBundle()` and the `CompositionBundle` interface deleted; the four style-index / indexer capabilities (`invalidateStyle`, `pushStyleFile`, `indexerReady`, `stopIndexer`) are now flat fields on `ProviderDeps`.
 - **Completion pipeline collapsed** — the two parallel pipelines added in v1.4.0 for `classnames/bind` vs `clsx / classnames` are unified into a single `findCompletionContext` helper that iterates once over bindings and style imports. `isInsideCxCall` renamed to `isInsideCall`. `detectClassUtilImports` moved to the binding-detector layer and exposed via `AnalysisEntry.classUtilNames`.
 - **Binding detector single-walk** — `collectImports` now makes exactly one pass over `sourceFile.statements` instead of two.
