+++ server/src/core/util/range-utils.ts
@@ -1,0 +1,9 @@
+import type { Range } from "@css-module-explainer/shared";
+
+export function rangeContains(range: Range, line: number, character: number): boolean {
+  const { start, end } = range;
+  if (line < start.line || line > end.line) return false;
+  if (line === start.line && character < start.character) return false;
+  if (line === end.line && character > end.character) return false;
+  return true;
+}
