+++ server/src/providers/scss-diagnostics.ts
@@ -2,6 +2,7 @@ import { DiagnosticSeverity, DiagnosticTag, type Diagnostic } from "vscode-langu
 import type { StyleDocumentHIR } from "../core/hir/style-types";
 import { findUnusedSelectors } from "../core/query/compute-unused-selectors";
 import type { SemanticWorkspaceReferenceIndex } from "../core/semantic/workspace-reference-index";
+import { toLspRange } from "./lsp-adapters";
 
 /**
  * Compute "unused selector" diagnostics for a single SCSS module file.
@@ -22,14 +23,4 @@ export function computeScssUnusedDiagnostics(
     message: `Selector '.${finding.canonicalName}' is declared but never used.`,
     tags: [DiagnosticTag.Unnecessary],
   }));
-}
-
-function toLspRange(range: {
-  readonly start: { readonly line: number; readonly character: number };
-  readonly end: { readonly line: number; readonly character: number };
-}): { start: { line: number; character: number }; end: { line: number; character: number } } {
-  return {
-    start: { line: range.start.line, character: range.start.character },
-    end: { line: range.end.line, character: range.end.character },
-  };
 }
